requirements.txt
matplotlib
pillow
numpy
pandas
wordcloud